requirements.txt
dearpygui==1.3.1
paho-mqtt==1.6.1
pyperclip==1.8.2
